tnns/services/betting-execution/requirements.txt
kafka-python==2.0.2
betfairlightweight==2.16.1
ciso8601==2.3.1
ujson==5.10.0
